requirements.txt
mysql-connector-python==8.2.0
pandas==2.1.4
requests==2.31.0
